FEAT: Login page
git commit -m "Add login form validation and error/success message display"

Add login form validation and error/success message display to login.controller.php.

The code now checks if the request method is POST and sanitizes the email and password inputs using the sanitizePhrase() function. It then checks if the email and password fields are empty. If either field is empty, an error message is appended to the $errorMessage variable. If the email is not a valid email format, another error message is added. Otherwise, a success message is added to the $successMessage variable.

The error and success messages are displayed using HTML div elements with appropriate CSS classes.

Also, require_once statements are added to include the functions.controller.php and login.view.php files.