This commit creates a new table called affiliated_companies.
Changes
Created a new table called affiliated_companies.
The table has the following columns:
id - An integer primary key.
company_name - A varchar(150) column for the company name.
phone - A varchar(45) column for the company phone number.
email - A varchar(150) column for the company email address.
address - A varchar(255) column for the company address.
The table uses the InnoDB engine.
Author
Ing. Nelson GE

Date
2023-07-27 15:42:32 PST

Notes
This table will be used to store information about affiliated companies.
The InnoDB engine is a good choice for this table because it is a transactional engine that provides good performance and reliability.
I hope this helps! Let me know if you have any other questions.